Modify some functions..

+++ 107590037_HW10/107590037_HW10/Mall.cpp
@@ -54,7 +54,7 @@ void Mall::AddOrderToCloth(int id)
 
 	¦^¶Ç­È: µL
 */
-void Mall::SelectCustomerAndShop(int customerIndex, int shopIndex)
+void Mall::SelectCustomerAndShop(size_t customerIndex, size_t shopIndex)
 {
     _shopIndex = shopIndex;
     _customerIndex = customerIndex;
